[API Integration Tests] usageApi service to expose the private telemetry API (#70057)
