+++ x-pack/test/reporting_api_integration/services.ts
@@ -10,6 +10,7 @@ import { filter, first, mapTo, switchMap, timeout } from 'rxjs/operators';
 // @ts-ignore no module definition
 import { indexTimestamp } from '../../plugins/reporting/server/lib/esqueue/helpers/index_timestamp';
 import { services as xpackServices } from '../functional/services';
+import { services as apiIntegrationServices } from '../api_integration/services';
 import { FtrProviderContext } from './ftr_provider_context';
 
 interface PDFAppCounts {
@@ -186,5 +187,6 @@ export function ReportingAPIProvider({ getService }: FtrProviderContext) {
 
 export const services = {
   ...xpackServices,
+  usageAPI: apiIntegrationServices.usageAPI,
   reportingAPI: ReportingAPIProvider,
 };
